Removed the src directory
